requirements.txt
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.1
cssselect2==0.7.0
Django==5.1.1
django-countries==7.6.1
django-filter==24.3
django-otp==1.5.4
django-otp-twilio==1.0.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
forex-python==1.8
html5lib==1.1
idna==3.8
inflection==0.5.1
kombu==5.4.0
lxml==5.3.0
oscrypto==1.3.0
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pypdf==4.3.1
pypng==0.20220715.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.16
